requirements.txt
asgiref==3.2.10
autopep8==1.5.3
Django==3.0.8
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
psycopg2==2.8.5
pycodestyle==2.6.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
toml==0.10.1
